refactor: consolidate shared Firebase APIs
